refactor: refactor the directory structure
