+++ src/main/java/gift/product/dto/ProductDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.product.dto;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
