+++ src/main/java/gift/user/controller/LoginPageController.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package gift.user.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
