+++ src/main/java/gift/user/utility/TokenUtility.java
@@ -5,7 +5,7 @@ import io.jsonwebtoken.security.Keys;
 import java.util.Date;
 
 // JWT 토큰을 생성해주는 utility 클래스
-public class SecurityUtility {
+public class TokenUtility {
 
     // 보안을 위해 token을 업데이트할 수 있도록 final로 선언하지 않기
     private static String secretKey = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
@@ -22,8 +22,6 @@ public class SecurityUtility {
             .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
             .compact();
     }
-
-    public static 
 
     // minute을 넣으면 밀리초로 반환하는 메서드
     private static long minuteToMillis(int minute) {
